+++ CHANGELOG.md
@@ -14,6 +14,9 @@
 # 5.12.0
 
 - feat: flutter error tracking support for web ([#243](https://github.com/PostHog/posthog-flutter/pull/243))
+
+- chore: improve survey color handling ([#233](https://github.com/PostHog/posthog-flutter/pull/233))
+
 - feat: add `userProperties` and `userPropertiesSetOnce` parameters to `capture()` method ([#254](https://github.com/PostHog/posthog-flutter/pull/254))
 
 # 5.11.1